+++ .eslintrc
@@ -9,6 +9,7 @@
   "parser": "babel-eslint",
   "rules": {
     "no-param-reassign": "off",
-    "import/prefer-default-export": "off"
+    "import/prefer-default-export": "off",
+    "no-prototype-builtins": "off"
   }
 }
