--- scripts/set_envs.sh
@@ -1,5 +1,0 @@
-#!/bin/bash
-
-#export all required env variables
-
-echo "Env vars was set"
