--- scripts/start.sh
@@ -1,6 +1,0 @@
-#!/bin/bash
-
-my_dir="$(dirname "$0")"
-
-source $my_dir/set_envs.sh
-node index.js
