+++ src/core/envconfig/typedefs.js
@@ -2,5 +2,4 @@
 
 export type Options = {
     prefix?: string,
-    allOptional?: boolean,
 }
