+++ src/core/mongodb/config.js
@@ -1,12 +1,17 @@
 // @flow
 
 class Config {
-  dbname: string = 'apki';
+  dbName: string = 'apki';
   host: string = '127.0.0.1';
   port: number = 27017;
   user: string = '';
   password: string = '';
-  poolsize: number = 5;
+  poolSize: number = 5;
 }
 
-export default Config;
+const cfg = new Config();
+
+export default cfg;
+export {
+  Config,
+};
